refactor(core): extract modules from common.py monolith
Extract focused modules from the 4000+ line common.py:
- duckdb_utils.py: DuckDB connection and query helpers
- remote.py: Remote file URL detection and error hints
- geometry_detection.py: Geometry column detection
- file_utils.py: File path handling utilities
- crs_utils.py: CRS/projection utilities

Update all imports across 51 files to use new module paths.
No re-exports from common.py (breaking change, pre-v1.0).

Part of PR #364.
