+++ geoparquet_io/core/add/bbox_metadata.py
@@ -7,11 +7,7 @@ import click
 import pyarrow.parquet as pq
 
 from geoparquet_io.core.check_parquet_structure import get_compression_info, get_row_group_stats
-from geoparquet_io.core.common import (
-    check_bbox_structure,
-    get_parquet_metadata,
-    parse_geo_metadata,
-)
+from geoparquet_io.core.common import check_bbox_structure, get_parquet_metadata, parse_geo_metadata
 from geoparquet_io.core.file_utils import safe_file_url
 from geoparquet_io.core.geometry_detection import find_primary_geometry_column
 from geoparquet_io.core.logging_config import debug, error, success
