+++ geoparquet_io/core/partition/admin_hierarchical.py
@@ -18,11 +18,11 @@ import duckdb
 from geoparquet_io.core.admin_datasets import AdminDatasetFactory
 from geoparquet_io.core.common import (
     check_bbox_structure,
-    find_primary_geometry_column,
     get_parquet_metadata,
-    safe_file_url,
     write_parquet_with_metadata,
 )
+from geoparquet_io.core.file_utils import safe_file_url
+from geoparquet_io.core.geometry_detection import find_primary_geometry_column
 from geoparquet_io.core.logging_config import debug, progress, success, warn
 from geoparquet_io.core.partition.common import (
     sanitize_filename,
