+++ geoparquet_io/core/streaming.py
@@ -228,7 +228,7 @@ def find_geometry_column_from_table(table: pa.Table) -> str | None:
         return geom_col
 
     # Fall back to common names
-    from geoparquet_io.core.common import STANDARD_GEOMETRY_NAMES
+    from geoparquet_io.core.geometry_detection import STANDARD_GEOMETRY_NAMES
 
     for name in STANDARD_GEOMETRY_NAMES:
         if name in table.column_names:
