+++ geoparquet_io/core/write_strategies/disk_rewrite.py
@@ -59,13 +59,12 @@ class DiskRewriteStrategy(BaseWriteStrategy):
     ) -> None:
         """Write query results to GeoParquet using DuckDB COPY then PyArrow rewrite."""
         from geoparquet_io.core.common import (
-            _wrap_query_with_wkb_conversion,
             compute_bbox_via_sql,
             compute_geometry_types_via_sql,
-            is_remote_url,
-            upload_if_remote,
             validate_compression_settings,
         )
+        from geoparquet_io.core.duckdb_utils import _wrap_query_with_wkb_conversion
+        from geoparquet_io.core.remote import is_remote_url, upload_if_remote
 
         configure_verbose(verbose)
         self._validate_output_path(output_path)
