+++ pyproject.toml
@@ -172,12 +172,15 @@ module = [
     "geoparquet_io.benchmarks.profile_report",
     "geoparquet_io.cli.decorators",
     "geoparquet_io.cli.main",
-    "geoparquet_io.core.add_bbox_column",
-    "geoparquet_io.core.add_country_codes",
-    "geoparquet_io.core.add_h3_column",
-    "geoparquet_io.core.add_kdtree_column",
-    "geoparquet_io.core.add_quadkey_column",
-    "geoparquet_io.core.add_s2_column",
+    "geoparquet_io.core.add.a5",
+    "geoparquet_io.core.add.admin_divisions",
+    "geoparquet_io.core.add.bbox",
+    "geoparquet_io.core.add.bbox_metadata",
+    "geoparquet_io.core.add.country_codes",
+    "geoparquet_io.core.add.h3",
+    "geoparquet_io.core.add.kdtree",
+    "geoparquet_io.core.add.quadkey",
+    "geoparquet_io.core.add.s2",
     "geoparquet_io.core.admin_datasets",
     "geoparquet_io.core.arcgis",
     "geoparquet_io.core.benchmark",
@@ -187,20 +190,29 @@ module = [
     "geoparquet_io.core.check_spatial_order",
     "geoparquet_io.core.common",
     "geoparquet_io.core.convert",
+    "geoparquet_io.core.crs_utils",
     "geoparquet_io.core.duckdb_metadata",
+    "geoparquet_io.core.duckdb_utils",
+    "geoparquet_io.core.file_utils",
+    "geoparquet_io.core.geo_metadata",
+    "geoparquet_io.core.geometry_detection",
     "geoparquet_io.core.extract",
     "geoparquet_io.core.extract_bigquery",
     "geoparquet_io.core.hilbert_order",
     "geoparquet_io.core.inspect_utils",
     "geoparquet_io.core.metadata_utils",
-    "geoparquet_io.core.partition_admin_hierarchical",
-    "geoparquet_io.core.partition_auto_resolution",
-    "geoparquet_io.core.partition_by_a5",
-    "geoparquet_io.core.partition_by_h3",
-    "geoparquet_io.core.partition_by_kdtree",
-    "geoparquet_io.core.partition_by_s2",
-    "geoparquet_io.core.partition_by_string",
-    "geoparquet_io.core.partition_reader",
+    "geoparquet_io.core.partition.admin_hierarchical",
+    "geoparquet_io.core.partition.auto_resolution",
+    "geoparquet_io.core.partition.by_a5",
+    "geoparquet_io.core.partition.by_h3",
+    "geoparquet_io.core.partition.by_kdtree",
+    "geoparquet_io.core.partition.by_quadkey",
+    "geoparquet_io.core.partition.by_s2",
+    "geoparquet_io.core.partition.by_string",
+    "geoparquet_io.core.partition.common",
+    "geoparquet_io.core.partition.reader",
+    "geoparquet_io.core.parquet_writer",
+    "geoparquet_io.core.remote",
     "geoparquet_io.core.reproject",
     "geoparquet_io.core.sort_by_column",
     "geoparquet_io.core.sort_quadkey",
