+++ tests/test_convert.py
@@ -27,13 +27,12 @@ from geoparquet_io.core.check_parquet_structure import (
     get_compression_info,
     get_row_group_stats,
 )
-from geoparquet_io.core.common import (
+from geoparquet_io.core.common import get_parquet_metadata, parse_geo_metadata
+from geoparquet_io.core.convert import convert_to_geoparquet
+from geoparquet_io.core.geometry_detection import (
     detect_parquet_geometry_column,
     find_primary_geometry_column,
-    get_parquet_metadata,
-    parse_geo_metadata,
 )
-from geoparquet_io.core.convert import convert_to_geoparquet
 
 
 def _check_all_passed(results: dict) -> bool:
